:lipstick: Converted homepage to a docs page
