+++ sites/custoplayer-docs/docs/components/_category_.json
@@ -1,8 +1,8 @@
 {
-    "label": "Components",
-    "position": 3,
-    "link": {
-        "type": "generated-index",
-        "description": "See the components below that you can put in an item container 👇🏽"
-    }
+  "label": "Components",
+  "position": 4,
+  "link": {
+    "type": "generated-index",
+    "description": "See the components below that you can put in an item container 👇🏽"
+  }
 }
