+++ sites/custoplayer-docs/docs/contributing.mdx
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 7
+sidebar_position: 8
 ---
 
 # Contributing
