+++ sites/custoplayer-docs/docs/homepage.mdx
@@ -1,0 +1,13 @@
+---
+sidebar_position: 1
+slug: /
+---
+
+import Homepage from '../src/components/Homepage';
+import Link from '@docusaurus/Link';
+
+# Custoplayer
+
+## A npm Package to Create Customizable Video Players
+
+<Homepage />
