+++ sites/custoplayer-docs/docs/quick-start.mdx
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 1
+sidebar_position: 2
 ---
 
 import CustoplayerExample from '../src/components/CustoplayerExample';
@@ -63,7 +63,7 @@ function CustoplayerExample() {
   return (
     <Custoplayer
       poster='https://custoplayer.nyc3.cdn.digitaloceanspaces.com/docs/custoplayer-demo-poster.png'
-      src='https://custoplayer.nyc3.cdn.digitaloceanspaces.com/docs/compressed-custoplayer-demo.mp4'
+      src='https://custoplayer.nyc3.cdn.digitaloceanspaces.com/docs/setting-up-video-qualities/custoplayer-demo-1080.mp4'
       values={{
         previewTooltip: {
           id: 'text',
