+++ sites/custoplayer-docs/docs/setting-up-subtitles.mdx
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 3
+sidebar_position: 6
 ---
 
 import SubtitlesExample from '../src/components/SubtitlesExample';
