+++ sites/custoplayer-docs/docs/setting-up-video-qualities.mdx
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 4
+sidebar_position: 7
 ---
 
 import QualityExample from '../src/components/QualityExample';
