+++ sites/custoplayer-docs/src/css/custom.css
@@ -90,33 +90,9 @@ nav {
   padding: 0.5rem !important;
 }
 
-header {
-  text-align: center;
-  padding: 2.5rem;
-}
-
-.hero-title {
-  font-weight: bold;
-  font-size: 2.65em !important;
-}
-.tagline {
-  font-weight: 600;
-  font-size: 1.8em;
-  line-height: 2.25rem;
-  margin: 1.35rem 0;
-}
-
-.home-main {
-  padding: 0.25rem 0 1.5rem 0;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
 .video-container {
-  width: min(90vw, 50rem);
-  margin: 0 auto;
+  width: 100%;
+  margin: 1rem auto 0.5rem;
 }
 
 label {
@@ -132,11 +108,12 @@ select {
   cursor: pointer;
 }
 
-@media screen and (max-width: 768px) {
-  .hero-title {
-    font-size: 2em;
-  }
-  .tagline {
-    font-size: 1.45em;
-  }
+.homepage-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+main {
+  padding: 1.25rem 1.5rem;
 }
